+++ packages/node/src/indexer/indexer.module.ts
@@ -11,6 +11,7 @@ import {
   StoreCacheService,
   WorkerDynamicDsService,
   PgMmrCacheService,
+  MmrQueryService,
 } from '@subql/node-core';
 import { SubqueryProject } from '../configure/SubqueryProject';
 import { ApiService } from './api.service';
@@ -56,6 +57,7 @@ import { WorkerUnfinalizedBlocksService } from './worker/worker.unfinalizedBlock
     PoiService,
     MmrService,
     PgMmrCacheService,
+    MmrQueryService,
     {
       provide: 'IProjectService',
       useClass: ProjectService,
@@ -71,6 +73,6 @@ import { WorkerUnfinalizedBlocksService } from './worker/worker.unfinalizedBlock
       },
     },
   ],
-  exports: [StoreService, MmrService],
+  exports: [StoreService, MmrService, MmrQueryService],
 })
 export class IndexerModule {}