+++ packages/node/src/indexer/indexer.manager.spec.ts
@@ -14,6 +14,7 @@ import {
   NodeConfig,
   ConnectionPoolService,
   StoreCacheService,
+  MmrQueryService,
 } from '@subql/node-core';
 import { Sequelize } from '@subql/x-sequelize';
 import { GraphQLSchema } from 'graphql';
@@ -158,7 +159,13 @@ function createIndexerManager(
     project,
   );
   const poiService = new PoiService(storeCache);
-  const mmrService = new MmrService(nodeConfig, storeCache, undefined);
+  const mmrService = new MmrService(
+    nodeConfig,
+    storeCache,
+    undefined,
+    undefined,
+  );
+  const mmrQueryService = new MmrQueryService(nodeConfig, sequilize);
   const unfinalizedBlocksService = new UnfinalizedBlocksService(
     apiService,
     nodeConfig,
@@ -175,6 +182,7 @@ function createIndexerManager(
     apiService,
     poiService,
     mmrService,
+    mmrQueryService,
     sequilize,
     project,
     storeService,