+++ packages/node/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Fixed
+- Sync with @node/core, various improvements for POI feature
 ### Changed
 - Update node-core and add `store-cache-upper-limit` flag (#144)