+++ packages/node/CHANGELOG.md
@@ -5,9 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Fixed
+- Fixed missing mmrQueryService in indexer module (#1885)
+
+## [2.9.2] - 2023-07-11
 ### Fixed
 - Handle RPC error for oversized block responses (#1876)
 - Modidy `TestingService` to use `TestRunner` from node-core (#1870)
+- Sync with @node/core, various improvements for POI feature (#1869)
 
 ## [2.9.1] - 2023-07-06
 ### Fixed
@@ -913,7 +919,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - bump @polkadot/api to 3.1.1
 
-[Unreleased]: https://github.com/subquery/subql/compare/node/2.9.1...HEAD
+[Unreleased]: https://github.com/subquery/subql/compare/node/2.9.2...HEAD
+[2.9.2]: https://github.com/subquery/subql/compare/node/2.9.1...node/2.9.2
 [2.9.1]: https://github.com/subquery/subql/compare/node/2.9.0...node/v2.9.1
 [2.9.0]: https://github.com/subquery/subql/compare/node/2.8.0...node/2.9.0
 [2.8.0]: https://github.com/subquery/subql/compare/node/2.7.0...node/2.8.0